backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-PyMongo==2.3.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.4
Flask-Limiter==3.3.1
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1
requests==2.31.0
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
lifelines==0.27.4
polyline==2.0.0
APScheduler==3.10.4
redis==5.0.0
python-engineio==4.5.1
python-socketio==5.8.0